discord-bot-tester/requirements.txt
discord.py>=2.3.2
PyYAML>=6.0
aiohttp>=3.9.0